Add LeafletTests folder
To run the leaflet stuff, start up a python server(using the command: python -m http.server) in the LeafletTests folder then browse in a web browser
to the stated address on the terminal, this is most likely http://0.0.0.0:8000/index.html.
This produces an interactable map which you can place markers/pins and polygons on then add comments in a popup.
When we start doing the django stuff, this folder will be removed and leaflet will be implimented in js ontop of django.